# pkg/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Tradebot is a tiny web api made with Go that takes array of stock prices and returns the best profit one could have made from 1 purchase and 1 sale using best practices of running microservices in Kubernetes.",
        "title": "Tradebot API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/praveensastry/tradebot"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/praveensastry/tradebot/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:9898",
    "basePath": "/",
    "paths": {
        "/delay/{seconds}": {
            "get": {
                "description": "waits for the specified period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chaos Engineering"
                ],
                "summary": "Delay",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delay wait in seconds",
                        "name": "seconds",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MapResponse"
                        }
                    }
                }
            }
        },
...
        "/trade/profit": {
            "post": {
                "description": "Calculates the Max profit from given prices for a share",
                "tags": [
                    "HTTP API"
                ],
                "summary": "Trade",
                "parameters": [
                    {
                        "description": "Array on prices",
                        "name": "prices",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MapResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.MapResponse": {
            "type": "object",
            "additionalProperties": {}
        }
    }
}